many: rework to make classic confinement a decision of the dirs module
The dirs module is really where this decision belongs to rather than
the release one. This decision depends on a global /snap directory
which does not exist when the global root dir is a different one.
