+++ release/release.go
@@ -67,16 +67,6 @@ func (o *OS) ForceDevMode() bool {
 	return false
 }
 
-// SupportsClassicConfinement returns true if the current distribution can run snaps using
-// classic confinement.
-func (o *OS) SupportsClassicConfinement() bool {
-	switch o.ID {
-	case "fedora", "rhel", "centos":
-		return false
-	}
-	return true
-}
-
 var (
 	osReleasePath         = "/etc/os-release"
 	fallbackOsReleasePath = "/usr/lib/os-release"
