+++ release/release_test.go
@@ -128,39 +128,3 @@ func (s *ReleaseTestSuite) TestForceDevMode(c *C) {
 		c.Assert(release.ReleaseInfo.ForceDevMode(), Equals, devmode, Commentf("wrong result for %#v", devmode))
 	}
 }
-
-func (s *ReleaseTestSuite) TestClassicSnapsAreEnabledOrDisabled(c *C) {
-	template := `NAME="###ID###"
-VERSION="0.1"
-ID="###ID###"
-ID_LIKE=###ID###
-PRETTY_NAME="###ID###"
-VERSION_ID="0.1"
-HOME_URL="http://###ID###.org/"
-SUPPORT_URL="http://###ID###.org/support/"
-BUG_REPORT_URL="https://###ID###.org/bugs/"`
-
-	for _, current := range []struct {
-		Name     string
-		Expected bool
-	}{
-		{"fedora", false},
-		{"rhel", false},
-		{"centos", false},
-		{"ubuntu", true},
-		{"debian", true},
-		{"suse", true},
-		{"yocto", true}} {
-		mockOSRelease := filepath.Join(c.MkDir(), fmt.Sprintf("mock-os-release-%s", current.Name))
-		dump := strings.Replace(template, "###ID###", current.Name, -1)
-		err := ioutil.WriteFile(mockOSRelease, []byte(dump), 0644)
-		c.Assert(err, IsNil)
-
-		reset := release.MockOSReleasePath(mockOSRelease)
-		defer reset()
-
-		os := release.ReadOSRelease()
-		c.Check(os.ID, Equals, current.Name)
-		c.Assert(os.SupportsClassicConfinement(), Equals, current.Expected)
-	}
-}
